Automatically find the name of the core bin file, instead of hardcoding
This is a PHP Experts, Inc., code donation. All rights and IP ownership will be
transferred to the host project and owning-organization/author(s) upon the merging
of this code into the parent project via this PR, when the PR is merged within
180 days of submission.

See https://github.com/PHPExperts-Contribs/ for more.

This code will revert to PHP Experts, Inc., ownership if the PR is not merged before
2025-06-25, or if the PR is closed and/or rejected without merging.

Fixes #109