Adds a Prebuilt command system + help command, which is run by default.
1. Adds a PrebuiltCommander.
2. Adds a prebuilt-command called "help" via the `PrebuiltCommands\HelpCommand`.
3. Modifies the `App::registerCommand()` method to `public function registerCommand(string $name, callable $callable, string $signature = '', ?string $description = null)`. This allows `HelpCommand` to generate help lists automatically.
4. Wrote code for creating pretty formatted help lists.
5. Wrote 100% integration test code coverage.
6. Updated the documentation.

This is a PHP Experts, Inc., conditional code donation. All rights and IP ownership will be
transferred to the host project and owning-organization/author(s) upon the merging of this code
into the parent project via this PR, when the PR is merged within 180 days of submission.

This code will revert to PHP Experts, Inc., ownership if the PR is not merged before 2025-06-25,
or if the PR is closed and/or rejected without merging.

The git commit acts as the ownership transfer when merged.